pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshling"
version = "0.1.0"
description = "Cross-platform terminal user interface for Meshtastic mesh networking devices"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Jascha Wanger / Tarnover, LLC"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Ham Radio",
    "Topic :: System :: Networking",
]
requires-python = ">=3.8"
dependencies = [
    "textual>=0.45.0",
    "meshtastic>=2.3.0",
    "pyserial>=3.5",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "asyncio-mqtt>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "bandit>=1.7.0",
    "mypy>=1.7.0",
    "textual-dev>=1.2.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
meshling = "meshling.main:main"

[project.urls]
Homepage = "https://github.com/meshling/meshling"
Repository = "https://github.com/meshling/meshling"
Issues = "https://github.com/meshling/meshling/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["meshling*"]

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101", "B601"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"